+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2026-02-28
+
+### Added
+
+- Added `prevent_repair` action to prevent items from being repaired in Anvils, Crafting, etc.
+
+### Fixed
+
+- Fixed boolean logic in filters.
+
 ## [1.4.5] - 2026-02-28
 
 ### Fixed